requirements.txt
flask
gunicorn
slackclient
